+++ R/get_CRU_df.R
@@ -70,7 +70,7 @@
 #'
 #' @seealso
 #' \code{\link{create_CRU_stack}}
-#' \code{\link{manage_CRU_cache}}
+#' \code{\link{CRU_cache}}
 #'
 #' @return A tidy data frame of CRU CL v. 2.0 climatology elements as a
 #' \code{\link[tibble]{tibble}} object