+++ tests/testthat/test_set_cache.R
@@ -1,14 +1,14 @@
 
 # test that .set_cache creates a cache directory if none exists ----------------
 
-test_that("test that set_cache creates a cache directory if none exists", {
+test_that("test that .set_cache creates a cache directory if none exists", {
   skip_on_cran()
-  unlink(rappdirs::user_cache_dir("getCRUCLdata"), recursive = TRUE)
+  unlink(CRU_cache$cache_path_get(), recursive = TRUE)
   cache <- TRUE
   .set_cache(cache)
-  expect_true(file.exists(file.path(rappdirs::user_cache_dir("getCRUCLdata"))))
+  expect_true(file.exists(CRU_cache$cache_path_get()))
   # cleanup
-  unlink(rappdirs::user_cache_dir("climcropr"), recursive = TRUE)
+  unlink(CRU_cache$cache_path_get(), recursive = TRUE)
 })
 
 # test that .set_cache does a cache directory if cache is FALSE ----------------